feat: (WIP) Landscape boundary mapbox migration
